+++ RemuxApp/Sources/SSH/RemuxCitadelSFTPClient.swift
@@ -416,12 +416,14 @@ struct RemuxSessionCitadelSFTPClientProvider: RemuxSFTPClientProvider {
     ) async throws -> RemuxSFTPClientLease<RemuxCitadelSFTPClient> {
         let registration = try await scope.begin()
         do {
+#if !REMUX_FILE_PROVIDER_EXTENSION
             let startedAt = GhosttyRuntimeTrace.latencyEnabled
                 ? GhosttyRuntimeTrace.nowNanos()
                 : nil
             GhosttyRuntimeTrace.latency(
                 "sftp.open begin host=\(hostDescription) source=session"
             )
+#endif
 
             // Citadel bounds subsystem negotiation internally. Await the raw
             // open so shutdown cannot release a shared root while a child is
@@ -429,11 +431,13 @@ struct RemuxSessionCitadelSFTPClientProvider: RemuxSFTPClientProvider {
             let sftp = try await SFTPClient.open(
                 overAuthenticatedSSHChannel: registration.rootChannel
             )
+#if !REMUX_FILE_PROVIDER_EXTENSION
             if let startedAt {
                 GhosttyRuntimeTrace.latency(
                     "sftp.open end host=\(hostDescription) source=session elapsed_ms=\(GhosttyRuntimeTrace.elapsedMilliseconds(from: startedAt))"
                 )
             }
+#endif
 
             let teardown = RemuxSFTPLeaseTeardown(
                 closeBorrowedChild: {
@@ -553,8 +557,10 @@ struct RemuxCitadelSFTPClientProvider: RemuxSFTPClientProvider {
         }
 
         do {
+#if !REMUX_FILE_PROVIDER_EXTENSION
             let sftpOpenStartedAt = GhosttyRuntimeTrace.latencyEnabled ? GhosttyRuntimeTrace.nowNanos() : nil
             GhosttyRuntimeTrace.latency("sftp.open begin host=\(rootConfiguration.host):\(rootConfiguration.port)")
+#endif
             let sftp = try await RemuxSFTPTimeout.run(
                 timeout: operationTimeout,
                 operation: {
@@ -568,11 +574,13 @@ struct RemuxCitadelSFTPClientProvider: RemuxSFTPClientProvider {
                     }
                 }
             )
+#if !REMUX_FILE_PROVIDER_EXTENSION
             if let sftpOpenStartedAt {
                 GhosttyRuntimeTrace.latency(
                     "sftp.open end host=\(rootConfiguration.host):\(rootConfiguration.port) elapsed_ms=\(GhosttyRuntimeTrace.elapsedMilliseconds(from: sftpOpenStartedAt))"
                 )
             }
+#endif
             let leaseState = RemuxSFTPLeaseTeardown(
                 closeChild: {
                     try await RemuxSFTPTimeout.run(
